\name{Biomod2ascii}
\alias{Biomod2ascii}
\title{Convert Biomod output files to ascii}
\description{Convert the output from BIOMOD (a dataframe with projections) to ASCII grids so they can be used in the MigClim simulation.}
\usage{Biomod2ascii (dataFrameOrFileName, XYcoord, ColumnsToConvert=NULL,
  projName="", inDir="", outDir="")}
\arguments{
  \item{dataFrameOrFileName}{Text missing...}
  \item{XYcoord}{Text missing...}
  \item{ColumnsToConvert}{Text missing...}
  \item{projName}{Text missing...}
  \item{inDir}{Text missing...}
  \item{outDir}{Text missing...}
}
