\name{MigClim.userGuide}
\alias{MigClim.userGuide}
\title{Displays the MigClim user guide in PDF format}
\description{Displays the MigClim user guide in PDF format. The function takes no argument.}
\usage{MigClim.userGuide()}
\value{The function does not return any value but opens the "MigClim_userGuide.pdf" PDF file containing the MigClim user guide.}
\seealso{MigClim.migrate ()}
